color picker

+++ DirectZobEngine/ZobObjects/ZobEntity.cpp
@@ -10,7 +10,7 @@ ZobEntity::ZobEntity(ZobType t, ZobSubType s)
 	m_type = t;
 	m_subType = s;
 	ulong id = GenerateId();
-	m_id = (u16)t * 1000 * 1000 * 1000 + (u16)s * 1000 * 1000 + id;
+	m_id = (u32)t * 1000 * 1000 * 1000 + (u32)s * 1000 * 1000 + (u32)id;
 	assert( m_entityMap.find(m_id) == m_entityMap.end());
 	ZobEntity::m_entityMap[m_id] = this;
 }
